ictu/x/credit/keeper/query_resume.go
package keeper

import (
	"context"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
	"ictu/x/credit/types"
)

func (k Keeper) Resume(goCtx context.Context, req *types.QueryGetResumeRequest) (*types.QueryGetResumeResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(goCtx)

	val, found := k.GetResume(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetResumeResponse{Resume: val}, nil
}
